reorganize the path of primitive apis
