+++ python/paddle/fluid/backward.py
@@ -485,7 +485,6 @@ def _addup_repetitive_outputs_(op_descs, block_idx, grad_var_to_var=None):
     And one op may yield its multiple outputs to the same variable.
     In these cases, the variable should be the accumulation of all the outputs.
     `sum_op`s are added to implement the accumulate.
-
     Args:
         grad_var_to_var(dict): used to build the mapping between grad var name and forward var name.
         Only for auto parallel.
@@ -656,14 +655,12 @@ def _find_not_need_ops(grad_op_descs, forward_ops, input_grad_names_set):
     The nodes of the computational graph composed of backward OPS should be
     interconnected. If there are unconnected sub-graphs in the computational graph,
     these sub-graphs should be cut off.
-
     Args:
         grad_op_descs(list[core.OpDesc]): The candidate backward OpDescs.
         forward_ops(list[Operator]): The forward ops.
         input_grad_names_set(set): this set is used to store the gradients' name
             which is generated by backward ops, and input_grad_names_set can help
             to prune the unnecessary backward ops.
-
     Return:
         (set[core.OpDesc]): A set of OpDescs which should be pruned.
     """
@@ -784,7 +781,6 @@ def _append_backward_ops_with_checkpoints_(
         block, ops, target_block, no_grad_dict, grad_to_var, checkpoints):
     """
     Create grad ops with forward ops, and insert them into given block
-
     Args:
         block(Block): the block where forward ops are
         ops(Op): the forward operators whose forward recomputation backward ops need to be added
@@ -793,7 +789,6 @@ def _append_backward_ops_with_checkpoints_(
             key(int) block index
             val(str): corresponding forward variable name
         checkpoints: variables that a user defined as checkpoint for forward recomputation
-
     Algorithms:
         0) deal with forward recomputing program descs
         1) find ops between checkpoints, i.e. recompute_segments
@@ -1013,7 +1008,6 @@ def _get_sub_block_path(sub_block,
     """
     Get output vars in subblock which will be assigned to parent block.
     It is used to find the grad path in subblock.
-
     Args:
         sub_block(Block): The sub-block in which to get op path.
         sub_block_op_desc: The op desc of the sub-block op such as 'while', 'conditional_block' and 'recurrent'.
@@ -1079,7 +1073,6 @@ def _append_backward_ops_(block,
                           rename_var_map=None):
     """
     Create all grad ops, and insert them into given block
-
     Args:
         block(Block): the block where forward ops are
         ops(Op): the forward operators whose backward ops need to be added
@@ -1297,7 +1290,6 @@ def _find_parent_op_(sub_block):
 def _append_backward_vars_(block, start_op_idx, grad_to_var, grad_info_map):
     """
     Create new variables required by backward pass.
-
     Args:
         block(Block): the block where new variables will be created
         start_op_idx(int): Only variables required by ops in block.ops[start_op_idx : ] will be created
@@ -1476,17 +1468,13 @@ def append_backward(loss,
                     distop_context=None):
     """
     :api_attr: Static Graph
-
     This function appends backward part to main_program.
-
     A complete neural network training is made up of forward and backward
     propagation. However, when we configure a network, we only need to
     specify its forward part. This function uses the chain rule to automatically
     generate the backward part according to the forward part.
-
     In most cases, users do not need to invoke this function manually.
     It will be automatically invoked by the optimizer's `minimize` function.
-
     Parameters:
         loss(Tensor): The loss Tensor of the network.
         parameter_list(list[Tensor|str]|tuple[Tensor|str], optional): List/Tuple of Parameters or Parameter.names
@@ -1523,56 +1511,42 @@ def append_backward(loss,
                                                gradient operator who has just
                                                triggered the callable object.
                                                Default: None.
-
     Returns:
         list of tuple ( :ref:`api_guide_tensor_en` , :ref:`api_guide_tensor_en` ): Pairs of parameter and its corresponding gradients.
         The key is the parameter and the value is gradient Tensor.
-
     Raises:
         AssertionError: If ``loss`` is not an instance of Tensor.
-
     Examples:
         .. code-block:: python
-
             import paddle
             import paddle.nn.functional as F
-
             paddle.enable_static()
-
             x = paddle.static.data(name='x', shape=[None, 13], dtype='int64')
             y = paddle.static.data(name='y', shape=[None, 1], dtype='float32')
             x_emb = paddle.static.nn.embedding(x, size=[100, 256])
             y_predict = paddle.static.nn.fc(x=x_emb, size=1, activation=None, name='my_fc')
             loss = F.square_error_cost(input=y_predict, label=y)
             avg_loss = paddle.mean(loss)
-
             # Get all weights in main_program, not include bias.
             all_weights = [param for param in paddle.static.default_main_program().block(0).all_parameters() if 'w_' in param.name]
             all_weights_name = [w.name for w in all_weights]
-
             # return all param_grads needed to be updated if parameter_list set default None.
             p_g_list1 = paddle.static.append_backward(loss=avg_loss)
             # output: [(embedding_0.w_0, embedding_0.w_0@GRAD), (my_fc.w_0, my_fc.w_0@GRAD), (my_fc.b_0, my_fc.b_0@GRAD)]
-
             # return the param_grads corresponding to parameter_list that can be list of param (Tensor).
             p_g_list2 = paddle.static.append_backward(loss=avg_loss, parameter_list=all_weights)
             # output: [(embedding_0.w_0, embedding_0.w_0@GRAD), (my_fc.w_0, my_fc.w_0@GRAD)]
-
             # parameter_list can be list of param.name (str).
             p_g_list3 = paddle.static.append_backward(loss=avg_loss, parameter_list=all_weights_name)
             # output: [(embedding_0.w_0, embedding_0.w_0@GRAD), (my_fc.w_0, my_fc.w_0@GRAD)]
-
             # no_grad_set can be set of Tensors that means grad will be cut off from these Tensors.
             p_g_list4 = paddle.static.append_backward(loss=avg_loss, no_grad_set=set([x_emb]))
             # output: [(my_fc.w_0, my_fc.w_0@GRAD), (my_fc.b_0, my_fc.b_0@GRAD)]
-
             # no_grad_set can be set of Tensor.name when the Tensor is created inside layers and can't be specified explicitly.
             p_g_list5 = paddle.static.append_backward(loss=avg_loss, no_grad_set=set(['my_fc.b_0']))
             # output: [(embedding_0.w_0, embedding_0.w_0@GRAD), (my_fc.w_0, my_fc.w_0@GRAD)]
-
             # return [] because all param_grads are filtered by no_grad_set.
             p_g_list6 = paddle.static.append_backward(loss=avg_loss, parameter_list=all_weights, no_grad_set=set(all_weights))
-
     """
     check_type(loss, 'loss', framework.Variable,
                'paddle.static.append_backward')
@@ -1855,7 +1829,6 @@ def _find_op_path_(block,
                    is_while=False):
     """
     It is used to find the grad path in `block`.
-
     Args:
         block(Block): The block in which to get op path.
         targets(list[Variable]): The target variables.
@@ -1931,7 +1904,6 @@ def _find_op_path_(block,
 def calc_gradient(targets, inputs, target_gradients=None, no_grad_set=None):
     """
     Backpropagate the gradients of targets to inputs.
-
     Args:
         targets(Tensor|list[Tensor]|tuple[Tensor]): The target Tensors
         inputs(Tensor|list[Tensor]|tuple[Tensor]): The input Tensors
@@ -1944,7 +1916,6 @@ def calc_gradient(targets, inputs, target_gradients=None, no_grad_set=None):
                                be automatically added into this set.
                                If this parameter is not None, the Tensors or Tensor.names in this set will be added to the default set.
                                Default: None.
-
     Return:
         (list[Tensor]): A list of gradients for inputs
         If an input does not affect targets, the corresponding gradient Tensor
@@ -2072,9 +2043,7 @@ def calc_gradient(targets, inputs, target_gradients=None, no_grad_set=None):
 @framework.static_only
 def gradients(targets, inputs, target_gradients=None, no_grad_set=None):
     """
-
     Backpropagate the gradients of targets to inputs.
-
     Args:
         targets (Tensor|list[Tensor]|tuple[Tensor]): The target Tensors.
         inputs (Tensor|list[Tensor]|tuple[Tensor]): The input Tensors.
@@ -2085,21 +2054,17 @@ def gradients(targets, inputs, target_gradients=None, no_grad_set=None):
             should be ignored. All Tensors with ``stop_gradient=True`` from all blocks will
             be automatically added into this set. If this parameter is not None, the Tensors or Tensor.names
             in this set will be added to the default set. Default: None.
-
     Return:
         (list[Tensor]): A list of gradients for inputs
         If an input does not affect targets, the corresponding gradient Tensor
         will be None.
-
     Examples:
     
         .. code-block:: python
           :name: code-example
             import paddle
             import paddle.nn.functional as F
-
             paddle.enable_static()
-
             x = paddle.static.data(name='x', shape=[None, 2, 8, 8], dtype='float32')
             x.stop_gradient=False
             y = paddle.static.nn.conv2d(x, 4, 1, bias_attr=False)
@@ -2114,10 +2079,9 @@ def gradients(targets, inputs, target_gradients=None, no_grad_set=None):
     check_type(target_gradients, 'target_gradients', (
         framework.Variable, list, tuple, type(None)), 'paddle.static.gradients')
 
-    from ..incubate.autograd.primx import _gradients
-    from ..incubate.autograd.utils import prim_enabled
-    if prim_enabled():
-        return _gradients(targets, inputs, target_gradients)
+    from paddle.incubate.autograd import primapi
+    if paddle.incubate.autograd.prim_enabled():
+        return primapi.gradients(targets, inputs, target_gradients)
 
     outs = calc_gradient(targets, inputs, target_gradients, no_grad_set)
     return _as_list(outs)
@@ -2127,9 +2091,7 @@ def gradients(targets, inputs, target_gradients=None, no_grad_set=None):
 def gradients_with_optimizer(program, optimizer, inputs=None, outputs=None):
     """
     :api_attr: Static Graph
-
     Backpropagate the gradients of the program and apply the gradients with the given optimizer.
-
     Args:
         program (Program): The input program.
         optimizer (Optimizer): The optimizer to apply the gradients.
@@ -2137,7 +2099,6 @@ def gradients_with_optimizer(program, optimizer, inputs=None, outputs=None):
             If None, the inputs will be created from the input variables in the given program. Default:None.
         outputs (Tensor|list[Tensor]|tuple[Tensor], optional): The output Tensors.
             If None, the outputs will be created from the output variables in the given program. Default: None.
-
     Return:
         tuple: tuple (optimize_ops, params_grads), A list of operators appended
             by gradients_with_optimizer and a list of (param, grad) variable pairs, param is
@@ -2145,21 +2106,16 @@ def gradients_with_optimizer(program, optimizer, inputs=None, outputs=None):
             The returned tuple can be passed to ``fetch_list`` in ``Executor.run()`` to
             indicate program pruning. If so, the program will be pruned by ``feed`` and
             ``fetch_list`` before run, see details in ``Executor``.
-
     Examples:
         .. code-block:: python
-
             import paddle
             import paddle.static as static
-
             paddle.enable_static()
-
             img = static.data(name='image', shape=[None, 784])
             pred = static.nn.fc(x=img, size=10, activation='relu')
             loss = paddle.mean(pred)
             opt_ops, pram_grads = paddle.fluid.backward.gradients_with_optimizer(static.default_main_program(), opt)
             print(opt_ops)
-
     """
     check_type(program, 'program', paddle.fluid.Program,
                'paddle.static.gradients_with_optimizer')
