+++ python/paddle/fluid/tests/unittests/autograd/CMakeLists.txt
@@ -8,4 +8,5 @@ endforeach(TEST_OP)
 
 set_tests_properties(test_autograd_functional_dynamic PROPERTIES TIMEOUT 160)
 set_tests_properties(test_autograd_functional_static PROPERTIES TIMEOUT 160)
-set_tests_properties(test_gradients_and_minimize PROPERTIES TIMEOUT 60)
+set_tests_properties(test_minimize PROPERTIES TIMEOUT 60)
+set_tests_properties(test_primapi PROPERTIES TIMEOUT 60)
