+++ python/paddle/fluid/tests/unittests/autograd/test_autograd_functional_static.py
@@ -18,13 +18,13 @@ import unittest
 import numpy as np
 import paddle
 import paddle.fluid as fluid
-from paddle.autograd.functional import forward_gradients
+from paddle.autograd.functional import _as_tensors
 
 import config
 import utils
 from utils import (_compute_numerical_batch_jacobian,
                    _compute_numerical_jacobian)
-from paddle.autograd.functional import _as_tensors
+from utils import gen_static_data_and_feed
 
 paddle.enable_static()
 
@@ -107,37 +107,6 @@ class TestVJPException(unittest.TestCase):
     def test_vjp(self):
         with self.assertRaises(self.expected_exception):
             self._vjp()
-
-
-def gen_static_data_and_feed(xs, v, stop_gradient=True):
-    feed = {}
-    if isinstance(xs, typing.Sequence):
-        static_xs = []
-        for i, x in enumerate(xs):
-            x = paddle.static.data(f"x{i}", x.shape, x.dtype)
-            x.stop_gradient = stop_gradient
-            static_xs.append(x)
-        feed.update({f'x{idx}': value for idx, value in enumerate(xs)})
-    else:
-        static_xs = paddle.static.data('x', xs.shape, xs.dtype)
-        static_xs.stop_gradient = stop_gradient
-        feed.update({'x': xs})
-
-    if isinstance(v, typing.Sequence):
-        static_v = []
-        for i, e in enumerate(v):
-            e = paddle.static.data(f'v{i}', e.shape, e.dtype)
-            e.stop_gradient = stop_gradient
-            static_v.append(e)
-        feed.update({f'v{i}': value for i, value in enumerate(v)})
-    elif v is not None:
-        static_v = paddle.static.data('v', v.shape, v.dtype)
-        static_v.stop_gradient = stop_gradient
-        feed.update({'v': v})
-    else:
-        static_v = v
-
-    return feed, static_xs, static_v
 
 
 def approx_jacobian(f, xs, dtype, eps=1e-5, batch=False):
@@ -510,58 +479,6 @@ class TestHessianFloat64(TestHessianFloat32):
             'second_order_grad').get('rtol')
         self.atol = config.TOLERANCE.get(self.dtype).get(
             'second_order_grad').get('atol')
-
-
-@utils.place(config.DEVICES)
-@utils.parameterize((utils.TEST_CASE_NAME, 'fun', 'xs', 'v'), (
-    ('matmul', paddle.matmul, (np.random.rand(2, 3), np.random.rand(3, 2)),
-     None), ('multiply', paddle.multiply,
-             (np.random.rand(2, 3), np.random.rand(2, 3)), None),
-    ('add', paddle.add, (np.random.rand(2, 3), np.random.rand(2, 3)),
-     None), ('input_not_sequence', paddle.tanh, np.random.rand(5, 5), None),
-    ('input_gradients_not_none', paddle.matmul,
-     (np.random.rand(3, 3), np.random.rand(3, 3)),
-     (np.random.rand(3, 3), np.random.rand(3, 3)))))
-class TestForwardGradients(unittest.TestCase):
-    def setUp(self):
-        self.dtype = str(self.xs[0].dtype) if isinstance(
-            self.xs, typing.Sequence) else str(self.xs.dtype)
-        self._rtol = config.TOLERANCE.get(str(self.dtype)).get(
-            "first_order_grad").get("rtol")
-        self._atol = config.TOLERANCE.get(str(self.dtype)).get(
-            "first_order_grad").get("atol")
-
-    def test_forward_gradients(self):
-        def expected():
-            sp = paddle.static.Program()
-            mp = paddle.static.Program()
-            with paddle.static.program_guard(mp, sp):
-                feed, static_xs, static_v = gen_static_data_and_feed(
-                    self.xs, self.v, stop_gradient=False)
-                _, ys_grad = paddle.autograd.jvp(self.fun, static_xs, static_v)
-            exe = paddle.static.Executor()
-            exe.run(sp)
-            return exe.run(mp, feed=feed, fetch_list=ys_grad)
-
-        def actual():
-            sp = paddle.static.Program()
-            mp = paddle.static.Program()
-            with paddle.static.program_guard(mp, sp):
-                feed, static_xs, static_v = gen_static_data_and_feed(
-                    self.xs, self.v, stop_gradient=False)
-                ys = self.fun(*static_xs) if isinstance(
-                    static_xs, typing.Sequence) else self.fun(static_xs)
-                ys_grad = forward_gradients(ys, static_xs, static_v)
-                paddle.incubate.autograd.prim2orig(mp.block(0))
-            exe = paddle.static.Executor()
-            exe.run(sp)
-            return exe.run(mp, feed=feed, fetch_list=ys_grad)
-
-        actual = actual()
-        expected = expected()
-        self.assertEqual(type(actual), type(expected))
-        np.testing.assert_allclose(
-            np.concatenate(actual), np.concatenate(expected))
 
 
 if __name__ == "__main__":
